Move HTML5 backend to C. Add missing display stubs
