+++ Backends/System/HTML5/Sources/kinc/backend/html5unit.c
@@ -1,0 +1,11 @@
+#include "audio.c.h"
+#include "display.c.h"
+#include "event.c.h"
+#include "mouse.c.h"
+#include "mutex.c.h"
+#include "semaphore.c.h"
+#include "system.c.h"
+#include "thread.c.h"
+#include "threadlocal.c.h"
+#include "video.c.h"
+#include "window.c.h"
